Reorganized files: resources and tests are now aside the associated code
