+++ src/fastoad/models/geometry/profiles/resources/airfoil_f_15_11.txt
