+++ src/fastoad/models/geometry/profiles/resources/airfoil_f_15_12.txt
