+++ src/fastoad/models/geometry/profiles/resources/airfoil_f_15_15.txt
