+++ src/fastoad/models/geometry/profiles/resources/naca23012.txt
