+++ src/fastoad/models/geometry/profiles/tests/__init__.py
@@ -1,6 +1,3 @@
-"""
-Management of wing profiles
-"""
 #  This file is part of FAST-OAD : A framework for rapid Overall Aircraft Design
 #  Copyright (C) 2020  ONERA & ISAE-SUPAERO
 #  FAST is free software: you can redistribute it and/or modify
@@ -13,5 +10,3 @@ Management of wing profiles
 #  GNU General Public License for more details.
 #  You should have received a copy of the GNU General Public License
 #  along with this program.  If not, see <https://www.gnu.org/licenses/>.
-
-from .get_profile import get_profile
